Restructure the full code base
now the codebase obeys the standard header-source file relation
